requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
transformers>=4.41.0
torch>=2.2.0
rank-bm25>=0.2.2
pandas>=2.2.0
numpy>=1.26.0
pinecone-client>=5.0.0
tqdm>=4.66.0
scikit-learn>=1.4.0
